+++ CHANGELOG.md
@@ -14,6 +14,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.7] - 2025-10-04
+**Added**
+- Check that will make sure blog post is valid before submitting to endpoint.
+
+**Fixed**
+- link to the generated blog post upon creation
+
+
 ## [0.0.6] - 2025-09-15
 **Fixed**
 - Github login not showing up.